backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
msal>=1.27.0
httpx>=0.27.0
netmiko>=4.3.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
python-jose>=3.3.0
